+++ src/app/config/app.constants.ts
@@ -11,20 +11,15 @@ export interface SymbolDetails {
 
 export const APP_CONSTS = {
   TITLE: 'Road To FIRE',
-  VERSION: '1.0.1',
+  VERSION: '1.0.2',
   // text to show when user first loads the current new version
-  WHATS_NEW_TITLE: `What's new in version 1.0.1`,
+  WHATS_NEW_TITLE: `What's new in version 1.0.2`,
   WHATS_NEW:
-    `Features:
-* Add action buttons on each asset-category tab of the account page, to quickly add/buy an asset.
-
-UI Changes:
-* Hide amount field for P2P and Real Estate, as it's irrelevant for these assets.
-* Allow editing of cash asset for bonds and P2P loans
-* Automatically fill cash asset field with default value on Cost, Dividend and Interest transaction dialogs.
+    `UI Changes:
+* Display tooltips of nearest point when hovering history charts, not only when hovering the exact point location.
 
 Bug Fixes:
-* Update previous payment date on bonds and P2P loans when manually adding an interest payment.`,
+* Correctly set currency symbol when a cash asset is selected by default in asset trade dialog`,
   REPOSITORY_URL: 'https://github.com/iLiviu/road-to-fire',
   STORAGE_API_KEYS: {
     DROPBOX: 'kbwh5504okeu3b9',